requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.9.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpie==3.2.4
httpx==0.28.1
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain-core==0.3.68
langchain-ollama==0.3.3
langsmith==0.4.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.3
narwhals==1.46.0
numpy==2.3.1
oauthlib==3.3.1
ollama==0.5.1
orjson==3.10.18
packaging==24.2
pandas==2.3.1
pillow==11.3.0
protobuf==6.31.1
psycopg2-binary==2.9.10
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
streamlit==1.46.1
stripe==12.3.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
whitenoise==6.9.0
zstandard==0.23.0
